requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pyTelegramBotAPI==4.15.4
Flask==3.0.1
python-dotenv==1.0.1
